# swagger.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: API Documentation
  description: This is a sample API documentation using Swagger
  termsOfService: http://example.com/terms/
  contact:
    name: API Support
    email: support@example.com
    url: http://example.com/support”
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: api.example.com
basePath: /v1
schemes:
  - https
  - http
paths:
  /vendors:
    get:
      tags:
        - vendors
      summary: Get all vendors
      operationId: vendors
      description: Returns a list of all vendors.
      responses:
        200:
          description: A list of vendors
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Vendor'
    post:
      summary: add new vendors
      tags:
        - vendors
      operationId: vendors
      description: Returns a list of all vendors.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vendor'
      responses:
        200:
          description: A list of vendors
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Vendor'
  /vendors/{id}:
    get:
      tags:
        - vendors
      summary: Get a vendor by ID
      operationId: vendors
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ID of the vendors to retrieve
      responses:
        200:
          description: vendors details
          schema:
            $ref: '#/definitions/Vendor'
        404:
          description: Vendor not found
    patch:
      tags:
        - vendors
      summary: update vendor by ID
      operationId: vendors
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ID of the vendors to updated
      responses:
        200:
          description: vendors succesfuly update
          schema:
            $ref: '#/definitions/Vendor'
        404:
          description: Vendor not found
    delete:
      tags:
        - vendors
      summary: delete a vendor by ID
      responses:
        200:
          description: vendors deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'deleted'
        404:
          description: User not found
...
components:
  schemas:
    Vendor:
      type: object
      properties:
        id:
          type: integer
          example: 12345
        name:
          type: string
          example: John
        surname:
          type: string
          example: Doe
        rang:
          type: integer
          example: 2
          min: 1
          max: 10
        collectionId:
          type: integer
          example: 1
      required:
        - id
        - firstName
        - lastName
...